Break index.js up into multiple feature modules
