+++ src/features/command/delete.js
@@ -1,7 +1,7 @@
 import isUrl from 'is-url'
-import config from '../config'
-import Post from '../models/post'
-import { getIdFromUrl, getPreview, isUserInChannel, sendEphemeralMessage, sendMessage } from '../utils'
+import config from '../../config'
+import Post from '../../models/post'
+import { getIdFromUrl, getPreview, isUserInChannel, sendEphemeralMessage, sendMessage } from '../../utils'
 
 // /prox delete <post number|url> [hard]
 export default async (client, command, args) => {
