+++ src/features/index.js
@@ -1,0 +1,5 @@
+export { default as command } from './command'
+export { default as movementLog } from './movement_log'
+export { default as postReply } from './post_reply'
+export { default as postReview } from './post_review'
+export { default as postSubmit } from './post_submit'
